refactor: consolidate deployment instance logic and remove instance assembler
- Merged functionality from the removed `instance_assembler.py` into `deployment_instance.py` to streamline the Config→Instance conversion process.
- Enhanced the `DeploymentInstance` class with improved parameter resolution, instance tree construction, and connection management.
- Updated documentation to clarify the three-layer data flow and the orchestration of system assembly.
- Removed obsolete methods and improved error handling during system setup.

These changes improve the maintainability and clarity of the codebase, facilitating better management of deployment instances in the Autoware System Designer.
